Allow secret values to tell you where they're stored
Sometimes, we need to refer to secrets that aren't stored in Hydroplane's secret store when defining
a process. For example, if we want to pass secrets to a process running in ECS, we can't just
get_secret() those secrets when defining the ECS task, since that would leak the secrets into the
task's definition. Instead, we need to be able to say that we're referring to a secret value stored
elsewhere (AWS SSM Parameter Store or AWS Secrets Manager, for instance) so the runtime can handle
that fact appropriately.